+++ km_api/know_me/journal/urls.py
@@ -13,6 +13,11 @@ urlpatterns = [
         name='entry-detail'),
 
     url(
+        r'^journal-entries/(?P<pk>[0-9]+)/comments/$',
+        views.EntryCommentListView.as_view(),
+        name='entry-comment-list'),
+
+    url(
         r'^users/(?P<pk>[0-9]+)/journal-entries/$',
         views.EntryListView.as_view(),
         name='entry-list'),