+++ docs/changelog.rst
@@ -15,6 +15,7 @@ Features
   * :issue:`246`: Create model for journal entries.
   * :issue:`247`: Add endpoint for listing and creating journal entries.
   * :issue:`248`: Create model for comments on journal entries.
+  * :issue:`249`: Add endpoint for listing and creating comments on journal entries.
 
 Bug Fixes
   * :issue:`254`: Fix Ansible creating duplicate crontab entries.